+++ include/mruby_integration/models/gamepad.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "mruby.h"
+
+struct Gamepad
+{
+  int index;
+};
+
+void
+append_models_Gamepad(mrb_state*);